requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
chromadb
google-generativeai
python-dotenv